Create RCodeMaxMin
This R code is designed to assist in risk assessment and management, offering a practical tool for analyzing and visualizing the potential impacts and exposures associated with specific risk scenarios. It involves a series of steps, including data generation, simulation, and visualization.

Setting Parameters: At the beginning of the code, essential parameters are defined. These include the number of random simulations (n), the desired confidence level (confidence_level), and the lower and upper bounds (lower and upper) that represent the range of a particular risk parameter. Additionally, we specify the expected number of trials in a given time horizon (events) and the probability of each trial resulting in an incident (chances).

Quantile Calculation: The code calculates the lower and upper quantiles based on the specified confidence level, which are crucial for assessing risk in the log-transformed space.

Log Transformation: The lower and upper bounds are log-transformed to facilitate risk assessment, given that certain risk factors may exhibit a log-normal distribution.

True Mean and Standard Deviation: The true mean and standard deviation of the log-transformed values are calculated, which are essential statistical parameters for risk analysis.

Simulating Impacts: Random values are generated from a log-normal distribution to simulate the impacts of a risk factor. These simulated impact values are then plotted in a histogram to visualize their distribution.

Simulating Incidents: Using a binomial distribution, the code simulates the number of incidents that occur within the specified time horizon. The distribution of the simulated incident counts is also displayed in a histogram.

Calculating Exposure: The code calculates the exposure by multiplying the simulated impacts and the simulated incident counts. A histogram is generated to visualize the distribution of exposure values.

Summary Statistics: Finally, the code provides a summary of exposure, offering statistics such as mean, median, and quartiles. These statistics offer insights into the risk exposure under consideration.

This code and the associated visualizations can be a valuable resource for risk analysts and decision-makers, helping them better understand and manage risk scenarios by providing a clear overview of potential impacts and exposures.

Prof. Hernan Huwyler MBA CPA